Refactor so phong and standard inherit from base mesh shader
